use unsigned integers and size_t in the flac plugin
git-svn-id: https://svn.musicpd.org/mpd/trunk@7324 09075e82-0dd4-0310-85a5-a0d7c8717e4f

+++ src/inputPlugins/_flac_common.h
@@ -144,9 +144,9 @@ typedef size_t flac_read_status_size_t;
 
 typedef struct {
 	unsigned char chunk[FLAC_CHUNK_SIZE];
-	int chunk_length;
+	size_t chunk_length;
 	float time;
-	int bitRate;
+	unsigned int bitRate;
 	FLAC__uint64 position;
 	OutputBuffer *cb;
 	DecoderControl *dc;
